requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
email-validator==2.0.0
python-multipart==0.0.6
databases[asyncpg]==0.8.0
redis==5.0.1
sentry-sdk==1.11.1
boto3==1.28.30
boto3-stubs[s3]==1.28.30
psycopg2-binary==2.9.7
SQLAlchemy==1.4.49
GeoAlchemy2==0.14.1
alembic==1.12.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
pytest==7.4.1
pytest-asyncio==0.21.1
httpx==0.24.1
asgi-lifespan==2.1.0